Update build.sh fixing path. Fix google-financial url. Improve project (#10)
* Update build.sh fixing path. Fix google-financial url. Improve project
setting; all for stockticker project.

Also fixed package.json to use devDependencies instead of dependencies
as they are used only in development.

Fixed google-financial url to get rid of cross-site request problem.

Improved project setting by centralizing running script inside NPM.
There are two possible commands

* `npm start` - to build stockticker component (generate stock-ticker-es6.v.html and stock-ticker-es6.v.js), then finally start a local http server.
* `npm build` - build stockticker only

above command will utilize `build.sh` script which is fixed for path to
access `vulcanize` and `criper` command as installed via package
manager.

* Separate two test cases. Separate JS code from `stock-ticker` template.

Two test cases that are possible are `stock-ticker` and
`stock-ticker-raw`. As we separated JS code from `stock-ticker`
template, and also the name of class for both `stock-ticker` and
`stock-ticker-raw` are the same, thus this will introduce error of
duplicate class name usage.

Thus we separate them into two test cases with link to click at the
bottom to test each one.

* Rename stockticker (raw) as StockTickerRaw. Add test suites.

There's a problem when testing with Jasmine as it's duplicate name of
StockTicker. Thus we resolve this by renaming it to be different to each
other.

Added Jasmine test cases for both of element.
Testing with Polymer needs a little of effort. In this case, we involve
karma-polymer on NPM to do the work. Notice karma.conf.js file for more
info on configuration.

* Add README.md file/

Added information about project, and NPM commands that can be executed to start, build and test the project.

* Get rid of bower and use entirely npm instead